# docs/swagger.yaml
basePath: /api
definitions:
  objects.EventTypeInfo:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  objects.GetEventTypesListResp:
    properties:
      eventTypes:
        description: |-
          EventTypes - массив словарей, каждый элемент которого содержит:
          id - ID типа события
          name - название типа события
          description - описание типа события
        items:
          $ref: '#/definitions/objects.EventTypeInfo'
        type: array
    type: object
  objects.GetObjectsListReq:
    properties:
      dateFrom:
        type: string
      dateTo:
        type: string
      eventTypeIds:
        items:
          type: integer
        type: array
    type: object
  objects.GetObjectsListResp:
    properties:
      objects:
        description: |-
          Objects - массив словарей, каждый элемент которого содержит:
          id - ID объекта
          title - название объекта
          description - описание объекта
          coordinates - координаты объекта
          eventDate - дата события объекта
          eventType - тип объекта
          previewUrlImage - URL картинки превью объекта
        items:
          $ref: '#/definitions/objects.ObjectInfo'
        type: array
    type: object
  objects.ObjectInfo:
    properties:
      description:
        type: string
      eventDate:
        type: string
      eventType:
        type: integer
      id:
        type: integer
      previewUrlImage:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
  description: Серверное REST API исторической платформы, разработанное на Go с использованием
    Gin.
  title: Документация API исторической платформы
  version: "1.0"
paths:
  /objects/get-event-types-list:
    get:
      description: Метод для получения списка типов событий
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/objects.GetEventTypesListResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение списка типов событий
      tags:
      - Объекты
  /objects/get-object-data/{id}:
    get:
      description: Метод для получения данных о конкретном объекте (метки на карте)
      parameters:
      - description: ID получаемого объекта (метки на карте)
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/objects.ObjectInfo'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение данных для конкретного объекта
      tags:
      - Объекты
  /objects/get-objects-list:
    post:
      consumes:
      - application/json
      description: Метод для получения списка объектов с возможностью фильтрации по
        типам событий и датам
      parameters:
      - description: Параметры фильтрации (опционально)
        in: body
        name: request
        schema:
          $ref: '#/definitions/objects.GetObjectsListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/objects.GetObjectsListResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение списка объектов
      tags:
      - Объекты
swagger: "2.0"